Allow generic container to have size_t elements rather than uint32_t.
Cast to uint32_t in those use cases when we are sure that container
only contains some few elements.

+++ re2c/src/codegen/emit_action.cc
@@ -168,7 +168,7 @@ void emit_accept_binary (OutputFile & o, uint32_t ind, bool & readCh, const Stat
 
 void emit_accept (OutputFile & o, uint32_t ind, bool & readCh, const State * const s, const accept_t & accepts)
 {
-	const uint32_t accepts_size = accepts.size ();
+	const uint32_t accepts_size = static_cast<uint32_t> (accepts.size ());
 	if (accepts_size > 0)
 	{
 		if (!DFlag)
